Non-Looping Low Health Beep
You need to first build the game, then replace the original se_low_health.s with the one from this branch. Then, build the game again, and it'll work. It's a bit of a pain but it's the easiest way I've found to edit the loop. Credit to myself for this one, I made it :)