Key System Setting Changes
Introduced a LEVEL CAP system based on the code from Infused Emerald by AsparagusEduardo. The options are STRICT, HARSH, and RELAXED. If the CAP is reached, the EXP. gain is stopped. This includes RARE CANDY use, and POKéMON DAY-CARE CENTER growth.

Introduced a NO ITEM HEALS system to disable healing item use from the field, from battle, and from both as an alternative to disabling the POKéCENTER. Items in the MEDICINE POCKET no longer have the “USE” text choice when selected. The POKé FLUTE is disabled in battle, but still functions for waking up SNORLAX. The BERRY POUCH no longer has the “USE” or “GIVE” text choice, however EV changing BERRIES still function as usual.

Added another EXP. MODIFIER option, as I have a busy life and a 4x option suits me better at times. 0x, 1/2x, 1x, and 2x still remain.

A “SHINY CLAUSE ON” and “SHINY CLAUSE OFF” option is now in NUZLOCKE MODE. This was a real pain in the ass to deal with. If switched on, and a shiny POKéMON is the first encounter on a route, the following non-shiny POKéMON will still be classed as the “first” encounter (with the “1” sprite to reflect this). The subsequent non-shiny encounters will be fainted as usual. If a shiny POKéMON is encountered after the first non-shiny encounter, it will still be valid and not immediately fainted. Switching the clause off is NUZLOCKE MODE without the SHINY CLAUSE. Switching it to DISABLED is the base game without NUZLOCKE MODE.

A fair amount of help system text and formatting overhaul, to reflect the new changes and fit the new information into the pitiful amount of space allowed by the system.

I'm sure my code is absolute dogshit to look at, so sorry for harming your eyes. This is still being tested, updated, and generally messed with. Use at your own risk?

Credit to me, for a lot of this. And of course to Deokishisu, for the original KEY SYSTEM SETTINGS menu and rom hack in general.